Clipppy nightly

+++ czkawka_core/src/bad_extensions.rs
@@ -333,7 +333,7 @@ impl BadExtensions {
                         checking_method: CheckingMethod::None,
                         current_stage: 1,
                         max_stage: 1,
-                        entries_checked: atomic_file_counter.load(Ordering::Relaxed) as usize,
+                        entries_checked: atomic_file_counter.load(Ordering::Relaxed),
                         entries_to_check,
                     })
                     .unwrap();
