+++ czkawka_core/src/temporary.rs
@@ -151,7 +151,7 @@ impl Temporary {
                     .unbounded_send(ProgressData {
                         current_stage: 0,
                         max_stage: 0,
-                        files_checked: atomic_file_counter.load(Ordering::Relaxed) as usize,
+                        files_checked: atomic_file_counter.load(Ordering::Relaxed),
                     })
                     .unwrap();
                 if !progress_thread_run.load(Ordering::Relaxed) {
